requirements.txt
flask
flask-smorest
python-dotenv
Flask-SQLAlchemy
marshmallow
flask-jwt-extended
passlib
flask-migrate

